crowd v1.3

+++ contracts/mvc.sol
@@ -14,9 +14,9 @@ contract mvcCrowd {
         uint _crowdTime;
     }
     
-    uint8 public userCount;
+    uint16 public userCount;
     CrowdInfo[1000] public info;
-    mapping(address=>uint8 ) public userId;
+    mapping(address=>uint16 ) public userId;
     
     event addCount(address _addr, uint _account);
     event onairDrop(address _addr, uint _acounnt);
@@ -31,16 +31,20 @@ contract mvcCrowd {
         uint _amount = totalSupply*10/100;
         info[0] = CrowdInfo(owner, _amount, now);
         totalCrowd += info[0]._amount;
+        // userid = info(num) = userCount--
+        // userCount++ = 1
+        // userId[_owner] = 0 ?
         userId[_owner] = userCount++;
         emit addCount(_owner, userId[_owner]);
     }
-    
+    // 将剩余的kccToken提取到其他账户, 并且加入到info数组中(不需要付出任何代价)
+    // 2中token间的关系: 通过已有的kccToken, 替换成买票需要mvcToken
     function airDrop(address _to, uint _value) public returns (bool success) {
         require(!isFinished);
         require( msg.sender == owner);
         require( _to != msg.sender);
         if ( address(0) != _to && totalCrowd + _value > 0 && totalCrowd + _value <= totalSupply) {
-            uint8 id = userId[_to];
+            uint16 id = userId[_to];
             if ( id >0 && id < userCount ){
                 // update
                 info[id]._amount += _value;
@@ -67,8 +71,9 @@ contract mvcCrowd {
     }
     
     function CrowInfo(address _owner) view public returns (uint _amount, uint _crTime) {
-        uint8 id = userId[_owner];
-        // 基金会余额为0 ?
+        // userAddress -> userId -> info[userId]
+        uint16 id = userId[_owner];
+        // id=0: 未添加至userid的mapping中, 查询_owner不为owner
         if (id == 0 && _owner != owner) {
             return (0, 0)
         }
@@ -77,7 +82,7 @@ contract mvcCrowd {
         emit onCrowInfo(_owner, _amount, _crTime);
         return (_amount, _crTime);
     }
-    
+    // 根据每个基金会会员的所占股份, 进行分成
     function ticketBooking() payable public returns (bool success) {
         require(isFinished);
         require(msg.value >= 100);
